+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.3.14 2026-02-22
 ### Improvement
+- Lint ignore TRY401.
+- Moved OCPP cached_property from sc_option_state to ocpp_charger.
 
 ## v0.3.13 2026-02-18
 ### Improvement